Add ProjectAssignmentsTable component
The code changes add a new component called ProjectAssignmentsTable. This component is responsible for displaying a table of project assignments. It fetches the assignments from the server using an HTTP request and sorts them based on different properties. The component also includes a loading spinner while the data is being fetched.